+++ NEWS.md
@@ -3,6 +3,8 @@
 ## New features
 
 * Added support for the extended bi-objective BBOB functions (FIDs 56-92) in makeBiObjBBOBFunction
+* Added extra helper functions to extract all problem instance-specific data from MPM2 generator
+* Added R-based evaluation environment for MPM2 generator which greatly accelerates evaluation in multi-objective settings
 
 ## Bugfixes